+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/XMLTilesDefsFileCheck.java
@@ -1,0 +1,75 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.source.formatter.checks;
+
+import com.liferay.portal.kernel.util.Tuple;
+import com.liferay.source.formatter.ElementComparator;
+import com.liferay.source.formatter.SourceFormatterMessage;
+import com.liferay.source.formatter.checks.util.SourceUtil;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import org.dom4j.Document;
+import org.dom4j.Element;
+
+/**
+ * @author Hugo Huijser
+ */
+public class XMLTilesDefsFileCheck extends BaseFileCheck {
+
+	@Override
+	public Tuple process(String fileName, String absolutePath, String content)
+		throws Exception {
+
+		Set<SourceFormatterMessage> sourceFormatterMessages = new HashSet<>();
+
+		if (fileName.endsWith("/tiles-defs.xml")) {
+			_checkTilesDefsXML(sourceFormatterMessages, fileName, content);
+		}
+
+		return new Tuple(content, sourceFormatterMessages);
+	}
+
+	private void _checkTilesDefsXML(
+			Set<SourceFormatterMessage> sourceFormatterMessages,
+			String fileName, String content)
+		throws Exception {
+
+		Document document = SourceUtil.readXML(content);
+
+		checkElementOrder(
+			sourceFormatterMessages, fileName, document.getRootElement(),
+			"definition", null, new TilesDefinitionElementComparator());
+	}
+
+	private class TilesDefinitionElementComparator extends ElementComparator {
+
+		@Override
+		public int compare(
+			Element definitionElement1, Element definitionElement2) {
+
+			String definitionName1 = getElementName(definitionElement1);
+
+			if (definitionName1.equals("portlet")) {
+				return -1;
+			}
+
+			return super.compare(definitionElement1, definitionElement2);
+		}
+
+	}
+
+}
